ocean anomaly reference based on average of previous run
the ocean anomaly approach for PISM input values (temp, salt) is now based on a reference state computed as the average of a previous ocean run. It is also possible to specify an ocean anomaly reference file from a previous coupled run which can be used in case of a restart of the coupled system. The old approach of using the ocean state of the first coupling iteration as a reference is removed.