adding option to pass timeseries of ocean forcing to PISM
When the coupling time step is greater than a year, the time series of yearly ocean output data can be passed to PISM to account for the forcing variability within a coupling time step. If this option is not enabled, the ocean forcing is averaged over the coupling time step before passed to PISM like before.