adding 'ice sheet runoff mechanism to impact sea level'
In ocean/sea ice only configuration (no atm), the ocean model uses a normalisation of surface mass fluxes (precip-evap+river=0) to keep the mass of the ocean and sea ice system constant.  (In MOM5 this is done via the 'zero_net_water_coupler' flag in the 'ocean_sbc' namelist).
To avoid that changes in the ice sheet mass (and subsequently ice-to-ocean runoff) are being balanced out when provided to the ocean
via the river routing scheme, the ice sheet runoff can be split:
 (1) regular mass flux to ocean in steady state (average at end of ice sheet standalone spinup
 (2) anomaly to the steady state/spinup mean
While (1) is provided via the regular river runoff mechanism, (2) is given to MOM5 via the new implemented 'runoff_slc' data table entry (which adds it to the same runoff field, but excludes it from the pme+river=0 normalisation). With the 'runoff_slc' flag this split can be enabled. It might be reasonable to switch it off during spinup phase of coupled ice/ocean simulations.  The provided 'runoff_reference_file' is required to have one timestamp only.