computing ice to ocean reference runoff from PISM extra output
The PISM-to-MOM_processing.py script has been extended to also compute the ice to ocean reference runoff from PISM's extra diagnostic variable 'surface_accumulation_flux'. It is used as an approximation for the combined ice to ocean runoff (calving, basal melting and surface runoff) in equilibrium state of the ice sheet under the presently given surface forcing. In the ice-ocean coupling framework this reference is used to split the current ice to ocean runoff into (1) the reference runoff, which is not causing a change in the ocean's sea level and thus subject to the global ocean's surface normalisation 'precip-evap+runoff=0' and (2) the anomaly runoff to the reference, which causes a sea level increase or decrease.