initial ice to ocean runoff when start coupling
The MOM template uses a land to ocean runoff file, where Antarctica is zeroed out. In the coupled setup this is instead filled by the runoff from the ice model. However, in a newly coupled run (no restart from a previous coupled run) there is no Antarctic runoff given to the ocean model, as MOM runs before PISM.
Now implemented: initial ice to ocean runoff to be specified in the settings which is applied for the first ocean iteration, typically the ice to ocean runoff used for spinning up the standalone ocean model.